refactor(code-structure): move IO.ts to /main
