+++ index.ts
@@ -1,2 +1,2 @@
 /* tslint:disable:file-name-casing */
-export {IO} from './src/IO'
+export {IO} from './src/main/IO'
