+++ test/internals/Counter.ts
@@ -2,7 +2,7 @@
  * Created by tushar on 2019-04-15
  */
 
-import {IO} from '../../src/IO'
+import {IO} from '../../src/main/IO'
 
 export const Counter = (n: number = 0) => {
   let count = n
